+++ aifm/inc/internal/object.ipp
@@ -10,13 +10,13 @@ FORCE_INLINE Object::Object() {}
 
 FORCE_INLINE Object::Object(uint64_t addr) : addr_(addr) {}
 
-FORCE_INLINE Object::Object(uint64_t addr, uint8_t ds_id, uint16_t data_len,
+FORCE_INLINE Object::Object(uint64_t addr, uint8_t ds_id, uint32_t data_len,
                             uint8_t id_len, const uint8_t *id)
     : Object(addr) {
   init(ds_id, data_len, id_len, id);
 }
 
-FORCE_INLINE void Object::init(uint8_t ds_id, uint16_t data_len, uint8_t id_len,
+FORCE_INLINE void Object::init(uint8_t ds_id, uint32_t data_len, uint8_t id_len,
                                const uint8_t *id) {
   set_ds_id(ds_id);
   set_data_len(data_len);
@@ -55,13 +55,13 @@ FORCE_INLINE void Object::free() {
   *reinterpret_cast<uint8_t *>(addr_ + kPtrAddrPos + kPtrAddrSize - 1) = 0xFF;
 }
 
-FORCE_INLINE void Object::set_data_len(uint16_t data_len) {
+FORCE_INLINE void Object::set_data_len(uint32_t data_len) {
   auto *ptr = reinterpret_cast<uint16_t *>(addr_ + kDataLenPos);
   *ptr = data_len;
 }
 
-FORCE_INLINE uint16_t Object::get_data_len() const {
-  auto *ptr = reinterpret_cast<uint16_t *>(addr_ + kDataLenPos);
+FORCE_INLINE uint32_t Object::get_data_len() const {
+  auto *ptr = reinterpret_cast<uint32_t *>(addr_ + kDataLenPos);
   return *ptr;
 }
