+++ aifm/inc/server_hashtable.hpp
@@ -16,9 +16,9 @@ public:
   ServerHashTable(uint32_t param_len, uint8_t *params);
   ~ServerHashTable();
   void read_object(uint8_t obj_id_len, const uint8_t *obj_id,
-                   uint16_t *data_len, uint8_t *data_buf);
+                   uint32_t *data_len, uint8_t *data_buf);
   void write_object(uint8_t obj_id_len, const uint8_t *obj_id,
-                    uint16_t data_len, const uint8_t *data_buf);
+                    uint32_t data_len, const uint8_t *data_buf);
   bool remove_object(uint8_t obj_id_len, const uint8_t *obj_id);
   void compute(uint8_t opcode, uint16_t input_len, const uint8_t *input_buf,
                uint16_t *output_len, uint8_t *output_buf);
